examples/python/requirements.txt
aws-cdk-lib==2.212.0
constructs>=10.0.0,<11.0.0
pytest>=8.0.0
cognito-ses-domain
